Updated the print format slightly
Reran the benchmark on the example and performance.

The insertion example results barely fluctuated.
As I restructured the entirety of Bench, it did get slightly slower.  The majority of the time is spent processing the benchmark data, since upon benchmarking `function() end` we can see that it takes ~7e-4 seconds (which is ~an eigth of the time).