Updated format and reduced max test amount
Added more spacing for histograms and metrics.  As different systems have different definitions for "\t" I will have to modify it later.

Reduced max test amount from 10,000 to 8,192 to reduce memory usage and speed up tests.  For tests that reached 10,000 , the results did not change when the reduction was performed.  Another reduction may occur, to 4,096, if necessary.