Added comparison example
I do not like how the results are displayed for the comparison section.

It is unclear exactly what is going on.  I know that it is saying that the average time for Length Insert is 41% more than for Table Insert and that it took 4e-8 seconds longer.  But it isn't crystal clear immediately.