0 looks cleaner and formatting improved
I noticed that 0.000e0 looked rather unsettling, so I converted it to 0.

I also fixed the spacing issue, however I see that High is 0 and that 0 is displayed.  Looking at the exact value, I see that it is attempting to compare 2.999e-7 to 3e-7, which is a very small difference, but it is non-zero, so I will let this issue exist until I see a situation where it shouldn't persist.

Aspects like 5.000e0 may need to be converted to 5.  So I will look into that.