Altered print format for >0 values
The prior issue with High regarding 0 and 0% should be somewhat fixed, as percentages are now slightly more accurate.

Values above 0 no longer have e<blah> and are not a forced length.  The forced length will remain for values less than 0 for legibility.

The units are sometimes confusing, so I need to add (seconds) and (kB) as needed.