Reverted to 2048 and spotted bug w/ histogram
Changed to 2048.  I feel like I have spotted a bug in the results for Bench.  130k is much more than the estimated 12k from the histogram count.

I also think that maybe the memory histogram could be skipping out on alot of values.  I know that a lot are garbage collected, but the average being 65.89 shows that most of the values are 66.  Hmm.  So maybe I keep values equal to low / high but no outliers.