Added the ability to print the output
Only the output to the FIRST run is printed, and the output can be modified using `updateOutput`.

I will personally be using it, however I understand that some use cases need a table of outputs.  Whereas other use cases don't want to print the output at all, for example if you are testing a compressor and just want to see how fast it is.  Although, by using `updateOutput` the user could also benchmark the output size.

`bench` is now returning an object whose metatable points to the `print`, `compare`, and more functions.  The choice may be reverted later on, but it was mainly to potentially save memory in situations where `bench` is called many times.  I haven't verified that luau-analyze doesn't spit out an error as my setup doesn't support analysis yet.