Improved hist performance further
Hist now performs 50% faster than the prior version.  Bench now performs 8% faster.

The main reason for hist's improvement is trying to avoid calculating `box = (v - histLow) // histRange` which needs floor division.  So instead of calculating box then continuing if it is below 0 or above boxCount, we check if it is below histLow or above histHigh.  We also keep the last value so we can potentially skip box calculations for the value, since it already was performed.

There is potentially an issue with correctness regarding min / low being 0 for memory when it should NOT be 0.  It may be due to an issue in table.move or remove copies, as I have yet to verify that section's correctness.  It may also be real data where the garbage collector happened to gc and have the memory at exactly 0 (or between 0 and the regular value) multiple times.