Improved stats hist
The calculation for box has been removed, it prior was potentially run hundreds of times, however the new calculation is only ever run 8 times or less.  There is no guaranteed improvement for benchmarks that produce highly similar values, however for benchmarks of more complex functions this yielded a 19% speedup (it fluctuates, sometimes 3% sometimes 40%).

The slower the old variant is, the faster this variant will be.  However as noted above, the simpler benchmarks will yield similar results in performance, with the old being potentially faster.

There is no change in the bench benchmark, as we use a simple function, however after using a more complex function:
table.sort(tab) where tab is already sorted.  We reached that it is painful to compare bench versions when the tested function is slow.  I got that it was 5% faster in the end, but this result could be inaccurate so take it with a grain of salt.