Added floor and nothing example
I will probably add more examples as I feel the need to.

I was questioning which floor approach was the fastest, and it turns out the most intuitive one was.

I was also questioning whether `do ... end` has any impact, it doesn't, and whether `for i=1,1 do ... end` has any impact, it does if optimize 2 is not used.  The example also proves that general tests should be performed using different optimization levels / native if possible.