Added external modification of low/high percentile
Added a way for users to adjust the low / high values.  They are by default the 1st percentile and 99th percentile respectively.

After further benchmarking of simple functions, the ability to modify the low/high value to expand / contract the histograms is necessary to ensure proper understanding.  Benchmark table.create against {} (singly), you will discover that a majority of runs fall within the same range, however upon modifying DIST_PERC from 1 to 5 we can "zoom in" and understand the subtle differences.  Although a better approach would be to loop, such an approach is not viable for expensive functions.